delete generated files in Scheduler Clean Workflow

+++ src/main/java/cz/zcu/kiv/server/scheduler/Manager.java
@@ -43,7 +43,7 @@ public class Manager {
     }
 
     //New Job arrives
-    public long addJob(Job job) throws SQLException {
+    public synchronized long addJob(Job job) throws SQLException {
         //Add job to database as a pending job
         job = Jobs.addJob(job);
         //Check for thread allocation
@@ -61,6 +61,7 @@ public class Manager {
     //Thread is available, start execution
     public void startJob(Job job) throws SQLException {
         CURRENT_THREAD_COUNT++;
+        
         job.setStatus(Status.RUNNING);
         Jobs.updateJob(job);
         JobThread jobThread = new JobThread(job) {
@@ -78,14 +79,18 @@ public class Manager {
 
     //Indicate thread is available and schedule pending job
     public synchronized void endJob() {
+
         CURRENT_THREAD_COUNT--;
         try {
             //Get list of pending jobs from database
             List<Job> pendingJobs = Jobs.getJobsByStatus(Status.WAITING.name());
             if (!pendingJobs.isEmpty()) {
 
-                //check if thread is available and schedule job
-                startJob(pendingJobs.get(0));
+                // check if thread is available and schedule job
+                Job job = pendingJobs.get(0);
+                job.setStatus(Status.RUNNING);
+                Jobs.updateJob(job);
+                startJob(job);
             }
         } catch (SQLException e) {
             logger.error(e);
@@ -114,8 +119,14 @@ public class Manager {
     public void clearJobs(String email) throws SQLException {
         List<Job> jobs = Jobs.getNotRunningOrWaitingJobs(email);
         Jobs.clearJobs(email);
-        for (Job item : jobs) {
-            String file = item.getWorkflowOutputFile();
+
+        if (jobs == null || jobs.size() == 0) return;
+
+        // all the jobs belongs to the owner = this email
+        String outputFolder = GENERATED_FILES_FOLDER + email;
+        for(Job item : jobs) {
+            // delete all the jsonFiles(in WORKING_DIRECTORY ) the owner has
+            String file = item.getWorkflowOutputFile(); 
             if (item != null && file != null && !file.equals("")) {
                 // delete json file
                 FileUtils.deleteQuietly(new java.io.File(item.getWorkflowOutputFile()));
@@ -140,7 +151,15 @@ public class Manager {
                     }
                 }
             }
+            // delete all the generatedFiles(in GENERATED_FILES_FOLDER ) of this job the owner has
+            List<String> files = item.getGeneratedFilesList();
+            if ( files == null || files.size() == 0) continue;
+            for(String generatedFile: files){
+                logger.info("delete " + email + "'s " + generatedFile);
+                FileUtils.deleteQuietly(new File(outputFolder + File.separator + generatedFile ));
+            }
         }
+
     }
 
 
@@ -192,7 +211,8 @@ public class Manager {
 
             String[] temps = fileStr.split("/");
             String fileName = temps[temps.length - 1]; //Sun May 26 20:35:48 CST 2019_Add.json
-            fileName = fileName.substring(0, fileName.length()-5); //Sun May 26 20:35:48 CST 2019_Add
+            fileName = fileName.substring(0, fileName.length()-5);
+
 
             int i = fileName.indexOf("_");
             //int j = fileName.lastIndexOf("_");
